Moved receive and send methods to Server Reader
Moved the receiveData and sendData methods from the ChatClient class to the ServerReader class.
